imp: place AnyCodec under sov-celestia-cw + reomve now unused PrettySlice + clean-ups
