+++ clients/sov-celestia-cw/src/context/mod.rs
@@ -15,11 +15,9 @@ use ibc_core::host::types::path::{
 };
 use ibc_core::primitives::proto::{Any, Protobuf};
 use prost::Message;
-use sov_celestia_client::types::sovereign::AnyCodec;
 
-use crate::types::{
-    parse_height, ClientType, ContractError, GenesisMetadata, HeightTravel, MigrationPrefix,
-};
+use crate::types::{ClientType, ContractError, GenesisMetadata, HeightTravel, MigrationPrefix};
+use crate::utils::{parse_height, AnyCodec};
 
 type Checksum = Vec<u8>;
