+++ clients/sov-celestia-cw/src/types/mod.rs
@@ -1,11 +1,11 @@
 mod client_type;
 mod error;
+mod helper;
 mod msgs;
 mod response;
-mod utils;
 
 pub use client_type::*;
 pub use error::*;
+pub use helper::*;
 pub use msgs::*;
 pub use response::*;
-pub use utils::*;
