+++ clients/sov-celestia-cw/src/types/msgs.rs
@@ -1,5 +1,4 @@
 //! Contains the definition of the messages that can be sent to the CosmWasm contract.
-
 use std::str::FromStr;
 
 use cosmwasm_schema::cw_serde;
