+++ clients/sov-celestia-cw/src/utils/codec.rs
