--- clients/sov-types/src/helpers/mod.rs
@@ -1,5 +1,0 @@
-mod codec;
-mod pretty;
-
-pub use codec::*;
-pub use pretty::*;
