--- clients/sov-types/src/helpers/pretty.rs
@@ -1,18 +1,0 @@
-use core::fmt::{Display, Error as FmtError, Formatter};
-
-pub struct PrettySlice<'a, T>(pub &'a [T]);
-
-impl<'a, T: Display> Display for PrettySlice<'a, T> {
-    fn fmt(&self, f: &mut Formatter<'_>) -> Result<(), FmtError> {
-        write!(f, "[ ")?;
-        let mut vec_iterator = self.0.iter().peekable();
-        while let Some(element) = vec_iterator.next() {
-            write!(f, "{element}")?;
-            // If it is not the last element, add separator.
-            if vec_iterator.peek().is_some() {
-                write!(f, ", ")?;
-            }
-        }
-        write!(f, " ]")
-    }
-}
