+++ clients/sov-types/src/lib.rs
@@ -2,7 +2,6 @@ mod aggregated_proof;
 mod client_params;
 mod consensus_params;
 mod error;
-mod helpers;
 
 #[cfg(feature = "test-util")]
 pub use aggregated_proof::test_util::*;
@@ -15,7 +14,6 @@ pub use client_params::test_util::*;
 pub use client_params::{SovereignClientParams, UpgradePath};
 pub use consensus_params::SovereignConsensusParams;
 pub use error::*;
-pub use helpers::*;
 
 /// Re-exports Sovereign SDK light clients proto types from `sov-ibc-proto`
 /// crate.
